Mac: stdfwd wrapper (#6)
