+++ lib/Util/Tokenizer.hpp
@@ -24,7 +24,7 @@ class Tokenizer
         VecStr GetLinesStr(const EnjoLib::Str & strData, bool excludeFirstLine = false) const;
         VecStr GetLines(EnjoLib::Istream & is, bool excludeFirstLine = false) const;
         VecStr FilterLines(const VecStr & lines, char tagComment = '#') const;
-        //VecStr GetLines(stdfwd::istream & is, bool excludeFirstLine = false) const;
+        //VecStr GetLines(STDFWD::istream & is, bool excludeFirstLine = false) const;
         void WorkOnLines(const EnjoLib::Str & fileName, IWorksOnLine & worker, bool excludeFirstLine = false) const;
         void WorkOnLines(EnjoLib::Istream & is, IWorksOnLine & worker, bool excludeFirstLine = false) const;
     protected:
